fix: Specify custom temp directory for clamscan
clamscan uses a default temp directory for internal handling of zip and tar files, extracting files contained in a tar or zip file to the temp directory. In the case of the `scan` lambda function, this can fill up the limited available space on the lambda VM, which results in clamscan erroring out with `LibClamAV Error: cli_writen: write error: No space left on device`.

This change creates a custom temp directory in the EFS mount and then passes it to clamscan with the --tempdir option, and then cleans up the temp directory before existing.

Fixes #64