feat: support use of imported buckets as sources (#470)
Clone of #465 from my personal account so that I can allow changes from upstream.

Passing an existing bucket to the construct is now supported, but the user must explicitly set the acceptResponsibilityForUsingExistingBucket or an error is thrown. When using an existing bucket, the bucket policy is not added automatically, but the relevant statement can be retrieved with getPolicyStatementForBucket(bucket) and the user is responsible to add that statement to the bucket. That can be done in different ways, but choice and responsibility is left to the user. For example a CfnOutput containing the policy statement can be created to expose the statement, or the policy can be explicitly replaced in the CDK project.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*